refactor(broker): split telemetry.rs into directory layout (C41 of PR-3)
Pure mechanical move. telemetry.rs (1398 lines) becomes:

- telemetry/mod.rs (~110 lines): re-exports of TelemetryEvent /
  TelemetryHandle / TelemetryService / run_telemetry_service, plus the
  chain-handle registry (init_for_chain / telemetry() / noop_handle /
  now_unix / proof_type_label / sum_chain_handles /
  global_committed_count / global_pending_preflight_count) and the
  CHAIN_HANDLES / NOOP_HANDLE statics.
- telemetry/event.rs (~145 lines): TelemetryEvent enum (broker-internal
  events).
- telemetry/handle.rs (~176 lines): TelemetryHandle producer + record_*
  methods. Fields widened to pub(super) so mod.rs can read counters in
  sum_chain_handles().
- telemetry/state.rs (~136 lines): InFlightRequest, CommitmentInfo, and
  the build_request_evaluated helper. Items widened to pub(super) so
  service.rs can construct/read them across modules.
- telemetry/service.rs (~957 lines): TelemetryService struct + impl,
  run_telemetry_service driver, read_heartbeat_intervals helper, and
  the original test module. All six module constants consolidated here
  (REQUEST_HEARTBEAT_INTERVAL_SECS / BROKER_HEARTBEAT_INTERVAL_SECS /
  EVICTION_INTERVAL_SECS / STALE_ENTRY_SECS / HEARTBEAT_RETRY_COUNT /
  HEARTBEAT_RETRY_SLEEP_MS); none are referenced from outside service.rs.

External paths crate::telemetry::{init_for_chain, telemetry, TelemetryEvent,
TelemetryHandle, TelemetryService, run_telemetry_service} preserved.
Every service's telemetry(chain_id).record_*(...) call-site keeps working
unchanged.
